refactor: Tasks page
