+++ frontend/src/pages/tasks/details/index.tsx
@@ -1,0 +1,1 @@
+export { TaskDetails } from "./TaskDetails";
