+++ frontend/src/pages/tasks/views/board/index.tsx
@@ -1,0 +1,1 @@
+export { TasksBoardView } from "./TasksBoardView";
