+++ frontend/src/pages/tasks/views/calendar/index.tsx
@@ -1,0 +1,1 @@
+export { TasksCalendarView } from "./TasksCalendarView";
