+++ frontend/src/pages/tasks/views/list/index.tsx
@@ -1,0 +1,1 @@
+export { TasksListView } from "./TasksListView";
